refactor(session): isolate interaction recovery
